backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
transformers==4.45.0
torch==2.5.0
sentence-transformers==3.2.0
numpy==1.26.4
scikit-learn==1.5.2
python-multipart==0.0.12
aiosmtplib==3.0.2
twilio==9.3.0
